requirements.txt
gunicorn
django
psycopg2-binary
django-environ
djangorestframework
pika
celery
redis
channels_redis
django_redis